third_party/blink/web_tests/svg/css/buffered-rendering-expected.txt

Test that an SVG image accepts all buffered rendering values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle("buffered-rendering", "") is "auto"
PASS computedStyle("buffered-rendering", "auto") is "auto"
PASS computedStyle("buffered-rendering", "dynamic") is "dynamic"
PASS computedStyle("buffered-rendering", "static") is "static"
PASS computedStyle("buffered-rendering", "0") is "static"
PASS computedStyle("buffered-rendering", "1") is "static"
PASS computedStyle("buffered-rendering", "true") is "static"
PASS computedStyle("buffered-rendering", "dynamic") is "dynamic"
PASS computedStyle("buffered-rendering", "0") is "dynamic"
PASS computedStyle("buffered-rendering", "1") is "dynamic"
PASS computedStyle("buffered-rendering", "true") is "dynamic"
PASS successfullyParsed is true

TEST COMPLETE
